requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
Automat==22.10.0
autopep8==2.0.2
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.3
constantly==15.1.0
contourpy==1.0.7
coverage==7.2.5
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
dataclasses==0.6
DateTime==5.1
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.0
fonttools==4.39.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
itemadapter==0.8.0
itemloaders==1.0.6
jedi==0.18.2
jmespath==1.0.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
packaging==23.1
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
Protego==0.2.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.15.1
pymongo==4.3.3
pyOpenSSL==23.1.1
pyparsing==3.0.9
pytest==7.3.1
pytest-cov==4.0.0
python-dateutil==2.8.2
python-telegram-bot==20.2
pytz==2023.3
pyzmq==25.0.2
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
rfc3986==1.5.0
-e git+https://github.com/PauMatas/ScoutingPlanner.git@502e55350d9ccf71af2f41cfa74134810278a371#egg=scouting_planner
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tldextract==3.4.0
tomli==2.0.1
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.15
w3lib==2.1.1
wcwidth==0.2.6
zope.interface==6.0
